refactor : 깨지는 테스트 코드 해결
